fix warnings when all generated files have been deleted
