+++ tools/serverpod_cli/lib/analyzer.dart
@@ -1,14 +1,15 @@
 /// Tools for analyzing the protocol files of a Serverpod project.
 library analyzer;
 
-export 'src/analyzer/analyzer.dart' show ProtocolAnalyzer;
 export 'src/analyzer/protocol_definition.dart' show ProtocolDefinition;
-export 'src/config/config.dart' show GeneratorConfig;
-export 'src/analyzer/yaml/entity_analyzer.dart' show ProtocolEntityAnalyzer;
+export 'src/config/config.dart' show GeneratorConfig, PackageType;
+export 'src/analyzer/entities/entity_analyzer.dart' show ProtocolEntityAnalyzer;
+export 'src/analyzer/dart/endpoints_analyzer.dart'
+    show ProtocolEndpointsAnalyzer;
 export 'src/analyzer/code_analysis_collector.dart' show CodeAnalysisCollector;
 export 'package:source_span/source_span.dart'
     show SourceSpanException; // The collector uses them.
-export 'src/analyzer/yaml/definitions.dart'
+export 'src/analyzer/entities/definitions.dart'
     show
         ProtocolEntityDefinition,
         ProtocolClassDefinition,
