+++ tools/serverpod_cli/lib/src/analyzer/entities/definitions.dart
